% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_sarla}
\alias{fit_sarla}
\title{Fit a SARLA model}
\usage{
fit_sarla(data, chains = 4, iter = 2000, ...)
}
\arguments{
\item{data}{Stan data}

\item{chains}{Chains}

\item{iter}{Iterations}

\item{...}{Other arguments to pass to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}}
}
\description{
Fit a SARLA model
}
