% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cohort_ids.R
\name{make_cohort_ids}
\alias{make_cohort_ids}
\title{Makes a matrix populated with an id corresponding to the cohort of the
observation.}
\usage{
make_cohort_ids(Nages, Nyears)
}
\arguments{
\item{Nages}{integer, number of ages}

\item{Nyears}{integer, number of years}
}
\value{
a matrix with dimensions Nages x Nyears with the cohort ids.
}
\description{
Makes a matrix populated with an id corresponding to the cohort of the
observation.
}
