% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_fill_data.R
\name{plot_and_fill_data}
\alias{plot_and_fill_data}
\title{Plot the simulated data object and use it to populate the stan input list.}
\usage{
plot_and_fill_data(
  dat,
  year_effects = 1L,
  cohort_effects = 0L,
  init_effects = 1L,
  plot = FALSE
)
}
\arguments{
\item{dat}{is simulated data object}

\item{year_effects}{scalar, 1 if year effects are estimated}

\item{cohort_effects}{scalar, 1 if cohort effects are estimated}

\item{init_effects}{scalar, 1 if initial size effects are estimated}
}
\value{
stan_dat a list containing needed Stan slots
}
\description{
Plot the simulated data object and use it to populate the stan input list.
}
