% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_length_data.R
\name{process_length_data}
\alias{process_length_data}
\title{process_length_data
A function that takes the input raw data, selects a species' data from it, and filters and standardizes it.}
\usage{
process_length_data(
  data__,
  common_,
  sex_,
  years_ = NULL,
  survey_string,
  minimum_n
)
}
\arguments{
\item{data__}{A Warehouse data object containing survey data}

\item{common_}{Common name of the species you are interested in}

\item{sex_}{sex to look at, must be "M" or "F"}

\item{years_}{The year at which the survey type changes, this can be NULL if you want the data from all or one survey across all years}

\item{survey_string}{A string in the `project` field that specifies which survey}

\item{minimum_n}{The minimum sample size used to set the maximum age of fish observed}
}
\value{
the processed length data
}
\description{
process_length_data
A function that takes the input raw data, selects a species' data from it, and filters and standardizes it.
}
