% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarla-package.R
\docType{package}
\name{sarla-package}
\alias{sarla-package}
\alias{sarla}
\title{The 'sarla' package.}
\description{
An autogressive state-space model in Stan that can estimate a
cohort, annual, or initial size effect on length-at-age data from
fish.
}
