% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sarla_sim}
\alias{sarla_sim}
\title{Simulate a random walk with trends}
\usage{
sarla_sim(
  sigma_p = 0.2,
  sigma_o = 0.1,
  Nages = 7,
  Nyears = 22,
  beta = 0.5,
  sigma_p_X0 = sigma_p,
  gamma_y_sd = 0,
  delta_c_sd = 0,
  eta_c_sd = 0
)
}
\arguments{
\item{sigma_p}{double setting process error sd. 0.2 by default.}

\item{sigma_o}{double setting observation error sd. 0.1 by default.}

\item{Nages}{integer setting the number of ages. 7 by default.}

\item{Nyears}{integer setting the number of years. 22 by default.}

\item{beta}{double setting the degree of autocorrelation. 0.5 by default.}

\item{gamma_y_sd}{double setting the sd of year variation. Default is 0.}

\item{delta_c_sd}{double setting the sd of cohort effect variation. Defaults
to 0.}

\item{eta_c_sd}{double setting the sd of the initial size cohort effect.
Defaults to 0.}

\item{sigma_p_x0}{double setting the sd of variation in starting values.
Defaults to equivalent to \code{sigma_p}}
}
\value{
list of simulated data
}
\description{
Simulate a random walk with trends
}
