% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_base.r
\name{stan_base}
\alias{stan_base}
\title{Run the stan model}
\usage{
stan_base(input_data, ...)
}
\arguments{
\item{input_data}{double setting process error sd. 0.2 by default.}

\item{...}{additional arguments to \code{\link{stan}}}
}
\value{
returns the stan output
}
\description{
Run the stan model
}
