% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_sarla}
\alias{fit_sarla}
\title{Fit a SARLA model}
\usage{
fit_sarla(data, chains = 4, iter = 2000, adapt_delta = 0.95, ...)
}
\arguments{
\item{data}{Stan data}

\item{chains}{Chains}

\item{iter}{Iterations}

\item{adapt_delta}{Adapt delta argument for sampling}

\item{...}{Other arguments to pass to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}. See the \verb{$sampling()} method link from \code{\link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel()}}.}
}
\description{
Fit a SARLA model
}
