% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_data.R
\name{fit_stan_data}
\alias{fit_stan_data}
\title{fit_stan_data
A function that takes the input data and feeds it into the Stan sarla model}
\usage{
fit_stan_data(input_data, ...)
}
\arguments{
\item{input_data}{the stan data to fit}

\item{...}{Other arguments to pass to \code{\link[=plot_and_fill_data]{plot_and_fill_data()}}.}
}
\description{
fit_stan_data
A function that takes the input data and feeds it into the Stan sarla model
}
