% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_std_length.R
\name{get_std_length}
\alias{get_std_length}
\title{get_std_length
This function will process a dataset and return the mean, standard deviation, and count of length observations between the reference period spanning 1995 - 2010}
\usage{
get_std_length(data__)
}
\arguments{
\item{data__}{A data frame with columns years, age_years, and length_cm}
}
\value{
a matrix of mean, sd, and n of length at age, where the summary statistics are calculated from 1995 - 2010
}
\description{
get_std_length
This function will process a dataset and return the mean, standard deviation, and count of length observations between the reference period spanning 1995 - 2010
}
